organize files of policeman-thief
